Role based account memberships

+++ app/views/components/pageflow/admin/grouped_folder_list.rb
@@ -23,22 +23,22 @@ module Pageflow
       private
 
       def build_all_link
-        ul(:class => 'folders') do
+        ul(class: 'folders') do
           li('data-role' => 'all') do
             if @active_id
-              link_to(t('pageflow.admin.folders.all'), main_app.admin_entries_path, :class => 'name')
+              link_to(t('pageflow.admin.folders.all'), main_app.admin_entries_path, class: 'name')
             else
-              span(t('pageflow.admin.folders.all'), :class => 'name')
+              span(t('pageflow.admin.folders.all'), class: 'name')
             end
           end
         end
       end
 
       def build_folder_list_grouped_by_accounts
-        ul(:class => 'accounts') do
+        ul(class: 'accounts') do
           folder_accounts.each do |account|
             li('data-id' => account.id) do
-              h4(account.name)
+              h4(account.name, class: 'account-name')
               build_folder_list(folders_by_account(account))
             end
           end
@@ -46,37 +46,41 @@ module Pageflow
       end
 
       def build_folder_list(folders)
-        ul(:class => 'folders') do
+        ul(class: 'folders') do
           folders.each do |folder|
             li('data-id' => folder.id) do
               text_node(link_to_folder(folder))
-              text_node(edit_folder_link(folder))
-              text_node(delete_folder_link(folder))
+              text_node(edit_folder_link(folder)) if authorized?(:configure_folder_on,
+                                                                 folder.account)
+              text_node(delete_folder_link(folder)) if authorized?(:configure_folder_on,
+                                                                   folder.account)
             end
           end
         end
       end
 
       def link_to_folder(folder)
-        link_to_unless(folder.id == @active_id.to_i, folder.name, main_app.admin_entries_path(:folder_id => folder), :class => 'name') do
-          span(folder.name, :Class => 'name')
+        link_to_unless(folder.id == @active_id.to_i,
+                       folder.name,
+                       main_app.admin_entries_path(folder_id: folder), class: 'name') do
+          span(folder.name, Class: 'name')
         end
       end
 
       def edit_folder_link(folder)
         link_to(t('pageflow.admin.folders.edit'),
                 main_app.edit_admin_folder_path(folder),
-                :title => t('pageflow.admin.folders.edit'),
-                :class => 'edit_folder')
+                title: t('pageflow.admin.folders.edit'),
+                class: 'edit_folder')
       end
 
       def delete_folder_link(folder)
         link_to(t('pageflow.admin.folders.destroy'),
                 main_app.admin_folder_path(folder),
-                :method => :delete,
-                :title => t('pageflow.admin.folders.destroy'),
-                :class => 'delete',
-                :data => {:confirm => t('pageflow.admin.folders.confirm_destroy')})
+                method: :delete,
+                title: t('pageflow.admin.folders.destroy'),
+                class: 'delete',
+                data: {confirm: t('pageflow.admin.folders.confirm_destroy')})
       end
 
       def folders_by_account(account)
